+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 3.3.1
 ----------------
 - Fix wkhtmltopdf error on Windows: prepend file:/// in PDF.php
 - Fix PHP error removed s.*, select each student field in Export.php
+- Optimize _makeNextSchool & _makeTeachers functions in miscExport.fnc.php
 
 Changes in 3.3
 --------------